Results are published on a channel specific to the request.
This has the effect of making individual producers accessible via
a publish. If you want to have a bunch of publishers, this avoids
the redundant parsing and "callback miss"/potential error emission
that occur when all producers receive results on the same 
subscription name. It also means producers aren't interchangeable,
but unless the callbacks are in shared state they never were.

Another, potentially better way to implement this would be to
give each producer a unique id, and encode that in the request.
The consumer could then namespace the result publish using that id.
With this set up, one could avoid the constant sub/unsub involved
in this commit, and instead have producers effectively own a
particular subscription name.